Add arcPath
Ignore-this: 59366b6804e01305f359139ee65502f8

An alternative to adding Arc/Sector constructors is to provide
functions that create paths in these shapes, which can then be
used as a polygon or line. Unfortunately this means that we don't
have access to useful information such as the scale factor, which
makes calculating an acceptable number of steps hard, so it
is included as an argument.

darcs-hash:20111231031239-c1e27-c2fe38bfda2e42b9625663a479260cdc8f34ec87