fix addColors
The switch from RGBA to the smart constructor makeColor broke addColors,
because addColors begins by creating an illegal color before normalizing
its components. With RGBA, that was fine, but makeColor clamps the
illegal color down to a legal color, thereby breaking addColors.

One easy way to observe the bug is that the color "azure" was the same
as the color "cyan". It's not in between "blue" and "cyan", as desired.