Add a 'BitmapFormat' to allow alt representations
The GL RGBA order (in memory, byte order is A|B|G|R) costs an additional
O(N) memory copy when the source is in the other order.  Instead, we can
expose GL's ABGR order for a performance win.  This change immediatly
allows us to clean up Internals.Rendering.Bitmap (removing the 'reverse'
routine) and positively impacts other packages (ex: gloss-juicy and
friday-juicypixels).

We also add a `RowOrder` for the same reason.  JuicyPixels and Friday
both work on the Top-to-bottom row order while GL generally is the other
way around.  Handling these cases without an O(N) copy can be important.

WRT API design, it is hard to expose more power without just re-casting
the GL API or a shoddy version of it.  I think I have a decently
powerful few options that don't cause headache but am open to
alternatives.  One thought was a more "hack-friendly" method where we
would use the exact openGL values and just pass them through the Gloss
layer, wrapping them in newtypes or something with an identical
representation (for 'coerce'), but after some consideration that seems a
bit over-the-top and not in the Gloss spirit.